Cleaned up the main, moved cleaning function to metadata_util.py and implemented proposed changes
