+++ .env.sample
@@ -1,26 +1,129 @@
-# Embedding model for Qdrant
+# ============================================================================
+# SCEPA Application - Environment Configuration Sample
+# ============================================================================
+# Copy this file to .env and fill in your actual values
+# DO NOT commit the .env file to version control!
+#
+# For local development: Use http:// URIs with TLS disabled
+# For production: Use https:// URIs with TLS automatically enabled
+# ============================================================================
+
+
+# ============================================================================
+# FILE STORAGE
+# ============================================================================
+# Directory where PDFs are downloaded and processed
+PDF_PATH=./data/pdfs/
+
+
+# ============================================================================
+# OPENAI / LLM CONFIGURATION
+# ============================================================================
+# API host for LLM and embeddings (can be OpenAI, Nebius, or compatible)
 OPENAI_HOST=https://api.tokenfactory.nebius.com/v1/
+
+# API key for accessing the LLM service
 OPENAI_API_KEY=
+
+# Model for LLM summaries
+OPENAI_LLM_MODEL=google/gemma-2-9b-it-fast
+
+# Model used for text embeddings (generates vector representations)
+# Examples: text-embedding-3-small, Qwen/Qwen3-Embedding-8B, etc.
 OPENAI_EMBEDDING_MODEL=Qwen/Qwen3-Embedding-8B
 
-# TypeDB
+# Vector dimension size for embeddings (must match the model's output)
+# Common values: 384, 768, 1024, 1536, 3072, 4096
+EMBEDDING_VECTOR_SIZE=4096
+
+
+# ============================================================================
+# TYPEDB CONFIGURATION
+# ============================================================================
+# TypeDB server URI for knowledge graph storage
+# - Development: http://localhost:1729 (TLS disabled)
+# - Production: https://dbscepadev.mads-han.src.surf-hosted.nl:1729 (TLS enabled)
+# Note: TLS is automatically enabled/disabled based on https:// prefix
 TYPEDB_URI=https://dbscepadev.mads-han.src.surf-hosted.nl:1729
+
+# TypeDB database name
 TYPEDB_DATABASE=knowledgeplatform
+
+# Path to TypeDB schema file (TypeQL schema definition)
+TYPEDB_SCHEMA_PATH=./schema/knowledge_graph_schema.tql
+
+# TypeDB username for authentication
 TYPEDB_USER=admin
+
+# TypeDB password for authentication
 TYPEDB_PASSWORD=
 
-# Qdrant
-QDRANT_URL=https://dbscepadev.mads-han.src.surf-hosted.nl:6333/
+
+# ============================================================================
+# QDRANT CONFIGURATION
+# ============================================================================
+# Qdrant vector database URI
+# - Development: http://localhost:6333 (HTTPS disabled)
+# - Production: https://dbscepadev.mads-han.src.surf-hosted.nl:6333 (HTTPS enabled)
+# Note: HTTPS is automatically enabled/disabled based on https:// prefix
+# Note: Include trailing slash if using URL path (e.g., https://host:port/)
+QDRANT_URI=https://dbscepadev.mads-han.src.surf-hosted.nl:6333/
+
+# Qdrant collection name for storing document vectors
 QDRANT_COLLECTION=knowledge_base
-QDRANT_VECTOR_SIZE=4096
+
+# Qdrant API key for authentication (leave empty for local development)
 QDRANT_API_KEY=
 
-# Langfuse
+
+# ============================================================================
+# LANGFUSE CONFIGURATION (Optional - for LLM observability)
+# ============================================================================
+# Langfuse server host for LLM trace logging and monitoring
 LANGFUSE_HOST=https://scepadeploy.mads-han.src.surf-hosted.nl
+
+# Langfuse secret key for API authentication
 LANGFUSE_SECRET_KEY=
+
+# Langfuse public key for API authentication
 LANGFUSE_PUBLIC_KEY=
 
-# Zotero
+
+# ============================================================================
+# ZOTERO CONFIGURATION
+# ============================================================================
+# Zotero library ID for fetching document metadata
 ZOTERO_LIBRARY_ID=
+
+# Zotero API key for accessing the library
 ZOTERO_API_KEY=
+
+# Zotero collection ID for filtering documents
+# Find in Zotero library URL or settings
 ZOTERO_COLLECTION_ID=
+
+
+# ============================================================================
+# DEVELOPMENT VS PRODUCTION NOTES
+# ============================================================================
+#
+# LOCAL DEVELOPMENT:
+# ─────────────────
+# OPENAI_HOST=http://localhost:8000/v1/  (if running local LLM server)
+# TYPEDB_URI=http://localhost:1729       (TLS automatically disabled)
+# QDRANT_URI=http://localhost:6333       (HTTPS automatically disabled)
+#
+# PRODUCTION:
+# ──────────
+# OPENAI_HOST=https://api.openai.com/v1/ (or compatible provider)
+# TYPEDB_URI=https://db.example.com:1729 (TLS automatically enabled)
+# QDRANT_URI=https://qdrant.example.com:6333 (HTTPS automatically enabled)
+#
+# TLS/HTTPS AUTO-DETECTION:
+# ────────────────────────
+# The application automatically enables TLS/HTTPS based on the URI scheme:
+# - URIs starting with https:// → TLS/HTTPS enabled
+# - URIs starting with http://  → TLS/HTTPS disabled
+# This works for both TYPEDB_URI and QDRANT_URI
+#
+# ============================================================================
